Renamed controllers per convention
